Pi-Videogames-final

+++ api/src/models/Videogame.js
@@ -16,7 +16,7 @@ module.exports = (sequelize) => {
       // unique: true,
     },
     rating: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.FLOAT,
       allowNull: false,
       // unique: true,
     },
